Make customProperties work as expected
I know you had said that this package will be left as is but I was hoping that wouldn't include PR's....

When using customProperties as it stands now, the expected results would be that the image shown would be the image with the customProperties only. That is not the case and although you can set the image and the customProperties are added, they aren't respected when looking up the image for detail or Form views.

This just handles that particular case so that customProperties are expected.